- Code refactoring for delegates and extension methods.
